[REFACTORE]: Reorderd folder of services and moved into models and move actual services to the folder moved handles and function into auth
