+++ Codes/backend/functions/src/callables/auth/changeUserEmail.ts
@@ -1,5 +1,5 @@
 import { onCall } from 'firebase-functions/v2/https';
-import { changeUserEmailHandler } from '../handlers/changeUserEmailHandler';
+import { changeUserEmailHandler } from '../../handlers/changeUserEmailHandler';
 
 export const changeUserEmail = onCall(
   { maxInstances: 10 },
