+++ Codes/backend/functions/src/callables/auth/checkServiceStatus.ts
@@ -1,5 +1,5 @@
 import { onCall } from 'firebase-functions/v2/https';
-import { checkServiceStatusHandler } from '../handlers/checkServiceStatusHandler';
+import { checkServiceStatusHandler } from '../../handlers/checkServiceStatusHandler';
 
 export const checkServiceStatus = onCall(
   { maxInstances: 10 },
