+++ Codes/backend/functions/src/callables/auth/login.ts
@@ -1,4 +1,4 @@
 import { onCall } from 'firebase-functions/v2/https';
-import { loginHandler } from '../handlers/loginHandlers';
+import { loginHandler } from '../../handlers/loginHandlers';
 
 export const login = onCall({ maxInstances: 10 }, loginHandler);
