+++ Codes/backend/functions/src/callables/auth/logout.ts
@@ -1,4 +1,4 @@
 import { onCall } from 'firebase-functions/v2/https';
-import { logoutHandler } from '../handlers/logoutHandlers';
+import { logoutHandler } from '../../handlers/logoutHandlers';
 
 export const logout = onCall({ maxInstances: 10 }, logoutHandler);
