+++ Codes/backend/functions/src/callables/auth/registerUser.ts
@@ -1,4 +1,4 @@
 import { onCall } from 'firebase-functions/v2/https';
-import { registerUserHandler } from '../handlers/registerUserHandler';
+import { registerUserHandler } from '../../handlers/registerUserHandler';
 
 export const registerUser = onCall({ maxInstances: 10 }, registerUserHandler);
