+++ Codes/backend/functions/src/callables/auth/sendLoginAlert.ts
@@ -1,5 +1,5 @@
 import { onCall } from 'firebase-functions/v2/https';
-import { sendLoginAlertHandler } from '../handlers/sendLoginAlertHandler';
+import { sendLoginAlertHandler } from '../../handlers/sendLoginAlertHandler';
 
 export const sendLoginAlert = onCall(
   { maxInstances: 10 },
