+++ Codes/backend/functions/src/callables/auth/sendPasswordReset.ts
@@ -1,5 +1,5 @@
 import { onCall } from 'firebase-functions/v2/https';
-import { sendPasswordResetHandler } from '../handlers/sendPasswordResetHandler';
+import { sendPasswordResetHandler } from '../../handlers/sendPasswordResetHandler';
 
 export const sendPasswordReset = onCall(
   { maxInstances: 10 },
