+++ Codes/backend/functions/src/handlers/auth/__tests__/checkServiceStatusHandler.test.ts
@@ -1,4 +1,4 @@
-import { checkServiceStatusHandler } from '../checkServiceStatusHandler';
+import { checkServiceStatusHandler } from '../../checkServiceStatusHandler';
 import * as emailService from '../../utils/emailService';
 import { CallableRequest } from 'firebase-functions/v2/https';
