+++ Codes/backend/functions/src/handlers/auth/checkServiceStatusHandler.ts
