+++ Codes/backend/functions/src/index.ts
@@ -1,9 +1,9 @@
-import './utils/firebaseAdmin'; // ensure initializeApp() has run
-import { sendLoginAlert } from './callables/sendLoginAlert';
-import { sendPasswordReset } from './callables/sendPasswordReset';
-import { registerUser } from './callables/registerUser';
-import { changeUserEmail } from './callables/changeUserEmail';
-import { checkServiceStatus } from './callables/checkServiceStatus';
+import './services/firebaseAdmin'; // ensure initializeApp() has run
+import { sendLoginAlert } from './callables/auth/sendLoginAlert';
+import { sendPasswordReset } from './callables/auth/sendPasswordReset';
+import { registerUser } from './callables/auth/registerUser';
+import { changeUserEmail } from './callables/auth/changeUserEmail';
+import { checkServiceStatus } from './callables/auth/checkServiceStatus';
 
 // Export callables
 export {
